wkhoops/requirements.txt
asgiref==3.5.2
<<<<<<< HEAD
backports.zoneinfo==0.2.1
boto3==1.24.22
botocore==1.27.22
Django==4.0.5
django-cors-headers==3.13.0
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.2.0
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
=======
coverage==6.4.1
Django==3.2
djangorestframework==3.13.1
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
>>>>>>> b16ffb9ef1dd6a0562de1ef4cbeacd11a225155d
